move banner to base InteractiveShell class
